+++ components/bridge/service/api/src/main/java/com/flipperdevices/bridge/service/api/FlipperServiceApi.kt
@@ -4,6 +4,7 @@ import com.flipperdevices.bridge.api.manager.FlipperRequestApi
 import com.flipperdevices.bridge.api.manager.delegates.FlipperConnectionInformationApi
 import com.flipperdevices.bridge.api.manager.service.FlipperInformationApi
 import com.flipperdevices.bridge.api.manager.service.FlipperRpcInformationApi
+import com.flipperdevices.bridge.api.manager.service.FlipperVersionApi
 
 /**
  * Provides access to the API operation of the device
@@ -32,6 +33,11 @@ interface FlipperServiceApi {
      * Returns wrapper of request api which provide device info from flipper
      */
     val flipperRpcInformationApi: FlipperRpcInformationApi
+
+    /**
+     * Provides information about the version of the API on Flipper. Null if not received.
+     */
+    val flipperVersionApi: FlipperVersionApi
 
     suspend fun disconnect()