+++ components/core/ktx/src/main/java/com/flipperdevices/core/ktx/jre/NumberKtx.kt
@@ -1,5 +1,7 @@
 package com.flipperdevices.core.ktx.jre
 
+import kotlin.math.abs
+import kotlin.math.log10
 import kotlin.math.roundToInt
 
 private const val PERCENT_MAX = 100
@@ -23,4 +25,9 @@ fun Long.toIntSafe(): Int {
 fun Float.roundPercentToString(): String {
     val processedPercent = if (this > 1.0f) 1.0f else if (this < 0.0f) 0.0f else this
     return "${(processedPercent * PERCENT_MAX).roundToInt()}%"
+}
+
+fun Int.length() = when (this) {
+    0 -> 1
+    else -> log10(abs(toDouble())).toInt() + 1
 }