+++ components/core/ui/theme/src/main/kotlin/com/flipperdevices/core/ui/theme/models/FlipperTypography.kt
@@ -35,9 +35,9 @@ data class FlipperTypography(
 
     val updateButton40: TextStyle,
     val updateText40: TextStyle,
-
     val monoSpaceM14: TextStyle,
     val monoSpaceM10: TextStyle
+    val flipperAction: TextStyle
 )
 
 internal val robotoFamily = FontFamily(