+++ components/keyscreen/impl/src/main/java/com/flipperdevices/keyscreen/impl/tasks/CloseEmulateAppTaskHolder.kt
@@ -1,0 +1,19 @@
+package com.flipperdevices.keyscreen.impl.tasks
+
+import com.flipperdevices.bridge.service.api.provider.FlipperServiceProvider
+
+object CloseEmulateAppTaskHolder {
+    private var closeEmulateAppTask: CloseEmulateAppTask? = null
+
+    @Synchronized
+    fun closeEmulateApp(serviceProvider: FlipperServiceProvider) {
+        if (closeEmulateAppTask != null) {
+            return
+        }
+        val localCloseEmulateAppTask = CloseEmulateAppTask(serviceProvider)
+        closeEmulateAppTask = localCloseEmulateAppTask
+        localCloseEmulateAppTask.start(Unit) {
+            closeEmulateAppTask = null
+        }
+    }
+}