+++ components/nfceditor/impl/src/main/java/com/flipperdevices/nfceditor/impl/viewmodel/NfcEditorViewModel.kt
@@ -9,12 +9,15 @@ import com.flipperdevices.core.keyinputbus.KeyInputBus
 import com.flipperdevices.core.keyinputbus.KeyInputBusListener
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
+import com.flipperdevices.core.log.verbose
 import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.nfceditor.impl.di.NfcEditorComponent
 import com.flipperdevices.nfceditor.impl.model.NFC_CELL_MAX_CURSOR_INDEX
 import com.flipperdevices.nfceditor.impl.model.NfcEditorCell
 import com.flipperdevices.nfceditor.impl.model.NfcEditorCellLocation
 import com.flipperdevices.nfceditor.impl.model.NfcEditorCursor
+import com.flipperdevices.nfceditor.impl.model.NfcEditorLine
+import com.flipperdevices.nfceditor.impl.model.NfcEditorSector
 import com.flipperdevices.nfceditor.impl.model.NfcEditorState
 import javax.inject.Inject
 import javax.inject.Provider
@@ -31,51 +34,82 @@ class NfcEditorViewModel : LifecycleViewModel(), LogTagProvider, KeyInputBusList
         keyInputBusProvider.get().subscribe(this, this)
     }
 
+    private val textUpdaterHelper = TextUpdaterHelper()
+
+    @Suppress("MagicNumber")
     var nfcEditorState by mutableStateOf(
         NfcEditorState(
-            listOf(
-                "B6 69 03 36 8A 98 02 00 64 8F 76 14 51 10 37 11".split(" ")
-                    .map { NfcEditorCell(it) },
-                "B6 69 03 36 8A 98 02 00 64 8F 76 14 51 10 37 11".split(" ")
-                    .map { NfcEditorCell(it) }
-            ),
-            cursor = null
+            List(size = (256 / 4)) { sectorIndex ->
+                NfcEditorSector(
+                    List(4) { lineIndex ->
+                        NfcEditorLine(
+                            sectorIndex * 4 + lineIndex,
+                            cells = "B6 69 03 36 8A 98 02 00 64 8F 76 14 51 10 37 11".split(" ")
+                                .map { NfcEditorCell(it) }
+                        )
+                    }
+                )
+            }
         )
     )
         private set
 
-    var currentActiveCell by mutableStateOf<NfcEditorCellLocation?>(null)
+    var nfcEditorCursor by mutableStateOf<NfcEditorCursor?>(null)
+        private set
 
-    fun onChangeSelection(location: NfcEditorCellLocation, position: Int) {
-        val localNfcEditorState = nfcEditorState.cursor
-        if (localNfcEditorState != null &&
-            location == localNfcEditorState.location &&
-            position == localNfcEditorState.position
+    var currentActiveCell by mutableStateOf<NfcEditorCellLocation?>(null)
+        private set
+
+    fun onCellFocus(location: NfcEditorCellLocation, isFocused: Boolean) {
+        if (!isFocused) {
+            if (currentActiveCell == location) {
+                currentActiveCell = null
+            }
+            return
+        }
+
+        currentActiveCell = location
+        if (nfcEditorCursor?.location != location) {
+            nfcEditorCursor = NfcEditorCursor(location, position = 0)
+        }
+    }
+
+    fun onChangeText(newText: String, location: NfcEditorCellLocation, position: Int) {
+        val oldCursor = nfcEditorCursor
+        if (oldCursor != null &&
+            location == oldCursor.location &&
+            position == oldCursor.position
         ) {
             return
         }
 
+        if (currentActiveCell != location) {
+            return
+        }
+
         val newCursor = if (position >= NFC_CELL_MAX_CURSOR_INDEX) {
-            if (location.columnIndex < nfcEditorState.lines[location.lineIndex].lastIndex) {
-                // If it is not last column:
-                NfcEditorCursor(
-                    lineIndex = location.lineIndex,
-                    columnIndex = location.columnIndex + 1,
-                    position = 0
-                )
-            } else if (location.lineIndex < nfcEditorState.lines.lastIndex) {
-                // If it is not last line:
-                NfcEditorCursor(
-                    lineIndex = location.lineIndex + 1,
-                    columnIndex = 0,
-                    position = 0
-                )
-            } else NfcEditorCursor(location, min(position, NFC_CELL_MAX_CURSOR_INDEX.toInt()))
+            val newLocation = oldCursor?.location?.increment(nfcEditorState.sectors)
+            if (newLocation == null) {
+                NfcEditorCursor(location, min(position, NFC_CELL_MAX_CURSOR_INDEX.toInt()))
+            } else NfcEditorCursor(newLocation, position = 0)
         } else NfcEditorCursor(location, position)
-        nfcEditorState = nfcEditorState.copy(
-            cursor = newCursor
+
+        val processedText = textUpdaterHelper.getProcessedText(
+            originalText = nfcEditorState[location].content,
+            newText = newText,
+            oldPosition = if (oldCursor?.location == location) oldCursor.position else 0,
+            newPosition = position
         )
-        println("On change selection. Input $location and $position. Output: $newCursor")
+
+        nfcEditorState = nfcEditorState.copyWithChangedContent(
+            location,
+            processedText
+        )
+        nfcEditorCursor = newCursor
+        verbose {
+            "On change selection. Input $location and $position. " +
+                "Output: $newCursor with $processedText"
+        }
     }
 
     override fun onKeyEvent(keyEvent: KeyEvent) {
@@ -89,27 +123,37 @@ class NfcEditorViewModel : LifecycleViewModel(), LogTagProvider, KeyInputBusList
     }
 
     private fun onPressBack() {
-        val cursor = nfcEditorState.cursor ?: return
+        val cursor = nfcEditorCursor ?: return
 
         // Execute only if current selector on zero
         if (cursor.position > 0) {
             return
         }
 
-        val newCursor = if (cursor.location.columnIndex > 0) {
+        val newCursor = cursor.location.decrement(nfcEditorState.sectors)?.let {
             NfcEditorCursor(
-                lineIndex = cursor.location.lineIndex,
-                columnIndex = cursor.location.columnIndex - 1,
-                position = NFC_CELL_MAX_CURSOR_INDEX.toInt() - 1
+                it,
+                NFC_CELL_MAX_CURSOR_INDEX.toInt() - 1
             )
-        } else if (cursor.location.lineIndex > 0) {
-            val newLineIndex = cursor.location.lineIndex - 1
-            NfcEditorCursor(
-                lineIndex = newLineIndex,
-                columnIndex = nfcEditorState.lines[newLineIndex].lastIndex,
-                position = NFC_CELL_MAX_CURSOR_INDEX.toInt() - 1
+        }
+
+        if (newCursor != null) {
+            val location = newCursor.location
+            val currentCell = nfcEditorState[location]
+            val oldText = currentCell.content
+            val newText = textUpdaterHelper.getProcessedText(
+                originalText = oldText,
+                newText = oldText,
+                oldPosition = NFC_CELL_MAX_CURSOR_INDEX.toInt(),
+                newPosition = newCursor.position
             )
-        } else null
-        nfcEditorState = nfcEditorState.copy(cursor = newCursor)
+            nfcEditorState = nfcEditorState.copyWithChangedContent(
+                location,
+                content = newText
+            )
+        }
+
+        currentActiveCell = newCursor?.location
+        nfcEditorCursor = newCursor
     }
 }